+++ src/platform_impl/linux/wayland/window/shim.rs
@@ -80,6 +80,9 @@ pub enum WindowRequest {
     /// Enable IME on the given window.
     AllowIme(bool),
 
+    /// Mark the window as opaque.
+    Transparent(bool),
+
     /// Request Attention.
     ///
     /// `None` unsets the attention request.
@@ -154,6 +157,9 @@ pub struct WindowHandle {
     /// Allow IME events for that window.
     pub ime_allowed: Cell<bool>,
 
+    /// Wether the window is transparent.
+    pub transparent: Cell<bool>,
+
     /// Visible cursor or not.
     cursor_visible: Cell<bool>,
 
@@ -194,6 +200,7 @@ impl WindowHandle {
             pending_window_requests,
             cursor_icon: Cell::new(CursorIcon::Default),
             is_resizable: Cell::new(true),
+            transparent: Cell::new(false),
             cursor_grab_mode: Cell::new(CursorGrabMode::None),
             cursor_visible: Cell::new(true),
             pointers: Vec::new(),
@@ -349,6 +356,19 @@ impl WindowHandle {
         }
     }
 
+    pub fn set_transparent(&self, transparent: bool) {
+        self.transparent.set(transparent);
+        let surface = self.window.surface();
+        if transparent {
+            surface.set_opaque_region(None);
+        } else {
+            let region = self.compositor.create_region();
+            region.add(0, 0, i32::MAX, i32::MAX);
+            surface.set_opaque_region(Some(&region.detach()));
+            region.destroy();
+        }
+    }
+
     pub fn set_ime_allowed(&self, allowed: bool, event_sink: &mut EventSink) {
         if self.ime_allowed.get() == allowed {
             return;
@@ -451,6 +471,13 @@ pub fn handle_window_requests(winit_state: &mut WinitState) {
                 }
                 WindowRequest::Minimize => {
                     window_handle.window.set_minimized();
+                }
+                WindowRequest::Transparent(transparent) => {
+                    window_handle.set_transparent(transparent);
+
+                    // This requires surface commit.
+                    let window_request = window_user_requests.get_mut(window_id).unwrap();
+                    window_request.redraw_requested = true;
                 }
                 WindowRequest::Decorate(decorate) => {
                     let decorations = match decorate {