+++ CHANGELOG.md
@@ -53,6 +53,7 @@ And please only add new entries to the top of this list, right below the `# Unre
 - **Breaking:** Remove the unstable `xlib_xconnection()` function from the private interface.
 - Added Orbital support for Redox OS
 - On X11, added `drag_resize_window` method.
+- Added `Window::set_transparent` to provide a hint about transparency of the window on Wayland and macOS.
 
 # 0.27.5